+++ source/Led.h
@@ -1,6 +1,7 @@
 #ifndef LED_H
 #define LED_H
 
+#include <Arduino.h>
 #include "Light.h"
 #include "OutputDigital.h"
 
@@ -9,7 +10,7 @@ namespace devuino
     class Led : public OutputDigital, public Light
     {
       public:
-        Led(int pin, int bitresolution = 8);
+        Led(byte pin, int bitresolution = 8);
         ~Led();
 
         void brightness(int value) override;
@@ -17,7 +18,7 @@ namespace devuino
         void on() override;
 
       protected:
-        int pin;
+        byte pin;
     };
 }
